video: encoder: stop spamming the syslog with resource replacement messages
virtio-video typically expects that a resource will be imported once and
reused afterwards, but the encoder works with arbitrary buffers that it
needs to reimport every time. This results in a bunch of

    Replacing source resource with id X

and

    Replacing dest resource with id X

messages printed in the host's syslog every time a new buffer is
submitted.

Demote these messages to debug so they don't clutter the syslog.

TEST=arc.VideoEncodeAccel.h264_360p_i420_vm while looking at
/var/log/messages

Change-Id: I37f6d034aa91bf6c8ff3ee8cc0977cc3ef76975a
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3924767
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>