e2e_tests: Add Suspend/Resume related tests.
Add snapshot/restore tests for the VM.
This test currently suspends the VM, then sends a command to it.
This command is stored in the buffer of the serial device.
A snapshot is taken with the command stored.
The device is resumed, which then runs the command.
A restore is triggered using the image taken earlier, which then runs
the image again.

This test will be developed to include more VM changes as more devices
get snapshot.

Bug=b:232437513
Test=./tools/dev_container e2e_tests/run

Change-Id: If6b39bd9b0f1d2962748cf089959f57f65407ced
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3949685
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Elie Kheirallah <khei@google.com>
Reviewed-by: Frederick Mayle <fmayle@google.com>