Enable mingw64 builders on CQ, disable coverage.
This is awkward. The previous .star config had the builder only
on postsubmit, but the actual .cfg files were outdated and enabled
it on the CQ already. Luckily it's been green, we just need to
disable coverage collection for postsubmit to pass as well.

This should be prevented by a presubmit check, unfortunately we
cannot do so in a reliable way since we need to authenticate
with luci in order to validate configs, which won't work
in the docker container.

BUG=b:254316054
TEST=None

Change-Id: Ie46b91bc691c7d033ff882a976dc0462ecf57c70
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3967433
Auto-Submit: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Zihan Chen <zihanchen@google.com>
Commit-Queue: Zihan Chen <zihanchen@google.com>