crosvm: handle disk ID more cleanly
Since we have several command-line options for specifying disks (disk,
rwdisk, root, and rwroot), we add an increasing disk ID each time we
encounter one of these options so we can preserve the order of
declaration later.

This ID is currently handle as a tuple alongside the actual DiskOption,
which forces us to use the `from_str_fn` argh attribute and requires the
ID to be explicitly specified if we deserialize RunCommand.

Replace the tuple by a dedicated local structure with its own FromStr
and From<DiskOption> implementations that assign the ID transparently.
This allows us to stop using `from_str_fn` and will also make
deserialization behave as we want for supporting configuration files.

BUG=b:218223240
TEST=Guest Linux boots as expected with the `--rwroot` option.

Change-Id: I291c119e6e9c4d76b71a2d6982d1504a3d689160
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3970362
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>