crosvm: rename fields of RunCommand for easier deserialization
We want to support configuration files where configuration options will
have the same name as our command-line arguments. For fields of
RunCommand which name is not identical to the argument itself, this
would mean adding a new serde helper attribute to make the name match we
one we already explicitly specified to argh.

It is just simpler and consistent to have the fields named exactly after
their corresponding argument, so do this. Some fields need to be
reordered in order to satisfy the #[remain::sorted] requirement, but no
other change is done otherwise.

BUG=b:218223240
TEST=`cargo run --features all-x86_64 -- run --help |sort` yields
identical output before and after with CL.

Change-Id: I1bf9d1e6a443afd61dd6d8180da42e485699353e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3970366
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>